backend/requirements.txt
# ── Web Framework ────────────────────────────────────────────────
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
python-multipart>=0.0.9

# ── Database / ORM ───────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.0,<3.0.0
asyncpg>=0.29.0,<1.0.0
alembic>=1.13.0,<2.0.0
greenlet>=3.0.0

# ── Task Queue ───────────────────────────────────────────────────
celery[redis]>=5.3.0,<6.0.0
redis>=5.0.0,<6.0.0

# ── Validation / Settings ───────────────────────────────────────
pydantic>=2.7.0,<3.0.0
pydantic-settings>=2.2.0,<3.0.0
eval-type-backport>=0.2.0          # Pydantic X|None syntax on Python <3.10

# ── HTTP Client ──────────────────────────────────────────────────
httpx>=0.27.0,<1.0.0

# ── Document Parsing ────────────────────────────────────────────
PyMuPDF>=1.24.0,<2.0.0
beautifulsoup4>=4.12.0,<5.0.0
lxml>=5.0.0

# ── Tokenizer ────────────────────────────────────────────────────
tiktoken>=0.7.0,<1.0.0

# ── LLM / AI ────────────────────────────────────────────────────
openai>=1.30.0,<2.0.0

# ── Vector Database ──────────────────────────────────────────────
qdrant-client>=1.9.0,<2.0.0

# ── Azure SDKs ───────────────────────────────────────────────────
azure-storage-blob>=12.20.0,<13.0.0
azure-identity>=1.16.0,<2.0.0

# ── Observability ────────────────────────────────────────────────
structlog>=24.1.0,<25.0.0
opentelemetry-api>=1.24.0,<2.0.0
opentelemetry-sdk>=1.24.0,<2.0.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-httpx>=0.45b0
opentelemetry-instrumentation-sqlalchemy>=0.45b0
opentelemetry-exporter-otlp>=1.24.0,<2.0.0
azure-monitor-opentelemetry-exporter>=1.0.0b21

# ── Utilities ────────────────────────────────────────────────────
python-dotenv>=1.0.0
numpy>=1.26.0,<2.0.0
